refactor: move diretório web para dentro de infra
- Ajusta chamada a package web
