+++ cmd/main.go
@@ -6,9 +6,9 @@ import (
 	"github.com/HaroldoFV/desafio/configs"
 	_ "github.com/HaroldoFV/desafio/docs"
 	"github.com/HaroldoFV/desafio/internal/infra/database"
+	"github.com/HaroldoFV/desafio/internal/infra/web"
+	"github.com/HaroldoFV/desafio/internal/infra/web/webserver"
 	"github.com/HaroldoFV/desafio/internal/usecase"
-	"github.com/HaroldoFV/desafio/internal/web"
-	"github.com/HaroldoFV/desafio/internal/web/webserver"
 	httpSwagger "github.com/swaggo/http-swagger"
 	"net/http"
 	"os"
