+++ internal/infra/web/gyroscope_handler.go
