+++ internal/infra/web/photo_handler.go
